moved routes and button endpoints to routes folder
